mwifiex: increase the global limit up to 4 SSID
Firmware for SDIO (88W8997), which is used in Turris MOX SDIO addon [1],
allows up to 4 SSID. Unfortunately, driver (even in mainline kernel)
has a global limit for all Marvell cards up to 3 SSID.

Pali Rohár tested this patch and verified that the SDIO Wi-Fi addon works
with the 4 SSID. So, let's increase the global limit from 3 to 4.

Ideally, this patch should be done differently before sending
it to Linux kernel. It means that limit definition should be moved to
the card-specific structure.

[1] https://docs.turris.cz/hw/mox/addons/#wi-fi-sdio