ath11k: control thermal support via symbol
Currently, thermal support will get built if CONFIG_THERMAL is reachable,
however this is not suitable for OpenWrt as with ALL_KMODS being set to y
ATH11K_THERMAL wont get selected and so hwmon and thermal kmods wont get
pulled in resulting in a build-failure.

So, to avoid that, lets do what is already done for ath10k and add a
config symbol into backports for enabling thermal support.

Signed-off-by: Robert Marko <robimarko@gmail.com>